Update ethers rs version
It is currently very hard to use this sdk and ethers rs in tandem due to the version of ethers being very outdated and types mismatching. This PR updates the ethers version to the latest version.

Also, run cargo fmt, run cargo clippy, and ignore Cargo.lock as this is a library.