Add reconnect support in ws_manager (#63)
Hi, I created a branch of the rust SDK which supports internal reconnect
in ws_manager @
https://github.com/snow-avocado/hyperliquid-rust-sdk/tree/ws-reconnect .
I found this simpliest to do rather than external reconnect because of
the way channels are used by the API.

A few notes:

1. I can add a flag to disable automatic reconnect so behavior is
generally unchanged if preferred.
2. Due to the way the internal APIs work, it does not currently
resubscribe for orderUpdates or userEvents. This can be fixed, but is
not as trivial as the reconnect logic.
3. I had to update my changes due to other recent changes for cleanup
(which are good and appreciated!), so I'm testing now but has been
working well for me.