Reduce dependencies for library consumers (#67)
Hello!

Cargo has the built-in feature to put examples in a dedicated folder.
These executables have access to the additional [dev-dependencies],
which can therefore be removed from the lib's dependencies. In
particular, moving tokio with the "full" feature enabled to
[dev-dependencies] reduces the weight of the crate for consumers by
quite a lot.

I haven't changed the version of any dependency.