Add Builder Codes (#66)
I've added the initial builder codes approval code having referenced it
from the python sdk and I think I'm missing something in the
implementation related to my understanding of the builders code API.

When I tried running it initially I got a 422 and now I get a 403 but
that's probably a personal environment issue. But would be great to get
some feedback on the current implementation. Please ignore the info!
lines.