fix: Return current timestamp when nonce lags
The function `next_nonce()` always returns a unique nonce, by storing
the current timestamp in a static variable and incrementing it before
returning. However, this variable can lag during times of
inactivity. Though there is a check that handles it by updating it, it
ends with returning the old nonce, which fails the current call.