Migrate to alloy (#119)
Replaced `ethers` with `alloy` and updated minor dependencies.

The only failing test was related to the connection ID, which now
produces a different hash due to changes in the underlying struct. Since
the connection ID only needs to be unique, this change is acceptable.
The test was updated accordingly and now passes.